test: Factor out a test package
